refactor: use relative admin compat consumers
