+++ rustfs/src/admin/handlers/kms_dynamic.rs
@@ -14,8 +14,8 @@
 
 //! KMS dynamic configuration admin API handlers
 
+use super::storage_compat::{read_admin_config, save_admin_config};
 use crate::admin::auth::validate_admin_request;
-use crate::admin::handlers::storage_compat::{read_admin_config, save_admin_config};
 use crate::admin::router::{AdminOperation, Operation, S3Router};
 use crate::app::context::{resolve_kms_runtime_service_manager, resolve_object_store_handle};
 use crate::auth::{check_key_valid, get_session_token};
